Task # 15 Added Controller Tests

+++ src/main/java/pl/futurecollars/invoicing/controller/invoice/InvoiceApi.java
@@ -27,13 +27,13 @@ public interface InvoiceApi {
 
     @GetMapping(value = "/{id}")
     @ApiOperation(value = "Get invoice by id")
-    ResponseEntity<Invoice> getById(@PathVariable long id);
+    ResponseEntity<Invoice> getById(@PathVariable int id);
 
     @PutMapping(value = "/{id}")
     @ApiOperation(value = "Update invoice with given id")
-    ResponseEntity<Invoice> updateById(@PathVariable long id, @RequestBody Invoice invoice);
+    ResponseEntity<Invoice> updateById(@PathVariable int id, @RequestBody Invoice invoice);
 
     @DeleteMapping("/{id}")
     @ApiOperation(value = "Delete invoice with given id")
-    ResponseEntity<Invoice> deleteById(@PathVariable long id);
+    ResponseEntity<Invoice> deleteById(@PathVariable int id);
 }
